Extract Navbar, HeroBanner, Footer, and icons from HomePage
HomePage.jsx was 547 lines with 12 inline SVG icons, a Navbar, HeroBanner, and Footer all in one file. Extracted each into its own component file and moved icon definitions to src/components/icons/index.jsx for reuse.

Addresses review comment on PR #4.
